requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.4.2
click==8.1.7
tqdm==4.66.1

# API Clients
requests==2.31.0
arxiv==1.4.8
pymed==0.8.9
scholarly==1.7.11

# Content Processing
PyPDF2==3.0.1
pdfminer.six==20221105
beautifulsoup4==4.12.2
lxml==4.9.3

# NLP and Analysis
spacy==3.6.1
transformers==4.34.0
sentence-transformers==2.2.2
nltk==3.8.1
bertopic==0.15.0
scikit-learn==1.3.1
gensim==4.3.1

# Report Generation
jinja2==3.1.2
markdown==3.5
matplotlib==3.8.0
seaborn==0.13.0
networkx==3.1
pydot==1.4.2
weasyprint==60.1
tabulate==0.9.0

# Database
sqlalchemy==2.0.22
alembic==1.12.0
redis==5.0.1

# Changelog and Version Control
gitpython==3.1.40
semver==3.0.2

# Web API
fastapi==0.103.2
uvicorn==0.23.2
gunicorn==21.2.0
pydantic-settings==2.0.3

# Testing
pytest==7.4.2
pytest-cov==4.1.0
responses==0.23.3
pytest-mock==3.11.1

# Development
black==23.9.1
flake8==6.1.0
isort==5.12.0
mypy==1.5.1
